+++ flask_app/templates/predictor.html
@@ -117,8 +117,8 @@
         <label for="warranty">Warranty (months):</label>
         <input type="number" id="warranty" name="warranty" min="0" required><br>
 
-        <label for="webcam">Webcam:</label>
-        <select id="webcam" name="webcam" required>
+        <label for="webcam_resolution">Webcam:</label>
+        <select id="webcam_resolution" name="webcam_resolution" required>
             <option value="Yes">Yes</option>
             <option value="No">No</option>
         </select><br>
